Cleaned up more unused code
