+++ engine/engine/wscript
@@ -66,7 +66,7 @@ namespace dmEngineVersion
     operating_sys = build_util.get_target_os()
     architecture = build_util.get_target_architecture()
 
-    sound_lib = 'sound2'
+    sound_lib = 'sound'
     if operating_sys == 'osx':
         conf.env.append_value('LINKFLAGS', ['-framework', 'Cocoa', '-framework', 'OpenGL', '-framework', 'OpenAL', '-framework', 'AGL', '-framework', 'IOKit', '-framework', 'Carbon', '-framework', 'CoreVideo'])
     elif operating_sys == 'ios':
