+++ engine/sound/src/sound.cpp
@@ -7,7 +7,6 @@
 #include <dlib/time.h>
 
 #include "sound.h"
-#include "sound2.h"
 #include "sound_codec.h"
 #include "sound_private.h"
 
@@ -186,6 +185,19 @@ namespace dmSound
     SoundSystem* g_SoundSystem = 0;
 
     DeviceType* g_FirstDevice = 0;
+
+    void SetDefaultInitializeParams(InitializeParams* params)
+    {
+        memset(params, 0, sizeof(InitializeParams));
+        params->m_OutputDevice = "default";
+        params->m_MasterGain = 1.0f;
+        params->m_MaxSoundData = 128;
+        params->m_MaxSources = 16;
+        params->m_MaxBuffers = 32;
+        params->m_BufferSize = 12 * 4096;
+        params->m_FrameCount = 768;
+        params->m_MaxInstances = 256;
+    }
 
     Result RegisterDevice(struct DeviceType* device)
     {
