+++ engine/sound/src/wscript
@@ -7,16 +7,11 @@ def configure(conf):
 def build(bld):
     bld.add_subdirs('openal')
 
-    source      = 'devices/device_null.cpp sound_codec.cpp sound_decoder.cpp sound_common.cpp sound2.cpp'.split()
-    source_null = 'devices/device_null.cpp sound_codec.cpp sound_decoder.cpp sound_common.cpp sound_null.cpp sound_generic.cpp'.split()
-    decoders    = 'decoders/decoder_stb_vorbis.cpp stb_vorbis/stb_vorbis.c decoders/decoder_wav.cpp'.split()
-    tremolo     = 'decoders/decoder_tremolo.cpp'.split()
+    source  = 'devices/device_null.cpp sound_codec.cpp sound_decoder.cpp sound.cpp'.split()
+    source += 'decoders/decoder_stb_vorbis.cpp stb_vorbis/stb_vorbis.c decoders/decoder_wav.cpp'.split()
 
     if bld.env['PLATFORM'] not in ['js-web', 'win32']:
-        decoders += tremolo
-
-    source += decoders
-    source_null += decoders
+        source += 'decoders/decoder_tremolo.cpp'.split()
 
     defines, include = ('', '')
     if re.match('arm.*?android', bld.env['PLATFORM']):
@@ -47,21 +42,13 @@ def build(bld):
     sound = bld.new_task_gen(features = 'cc cxx cstaticlib ddf',
                              includes = '. %s' % include,
                              defines = defines,
-                             target = 'sound2',
+                             target = 'sound',
                              source = source)
-
-    # sound_null is only used for the headless build.
-    sound_null = bld.new_task_gen(features = 'cc cxx cstaticlib ddf',
-                                  includes = '.',
-                                  defines = defines,
-                                  target = 'sound_null',
-                                  source = source_null)
 
     bld.install_files('${PREFIX}/include/sound', 'sound.h')
 
     if sys.platform == 'win32':
         #NOTE: _XBOX to get static lib and avoid dllimport/dllexport stuff
         sound.defines = '_XBOX'
-        sound_null.defines = '_XBOX'
 
     bld.add_subdirs('test')
